feat(next-web): Add new modules and improve API documentation integration
新增 `next-web-api-doc`、`next-web-actuator` 和 `next-web-pay` 模块以支持 API 文档生成、应用监控和支付功能。更新 `Cargo.toml` 文件以包含新模块，并在 `next-web` 中增加 `enable-api-doc` 特性以启用 API 文档功能。

- 在 `Cargo.toml` 中添加 `next-web-api-doc`、`next-web-actuator` 和 `next-web-pay` 模块
- 在 `Cargo.toml` 中添加 `utoipa-swagger-ui` 依赖以支持 Swagger UI
- 创建 `next-web-api-doc` 模块，包含生成 OpenAPI 文档的代码
- 创建 `next-web-actuator` 模块，包含应用监控相关的代码
- 创建 `next-web-pay` 模块，提供支付功能
- 更新 `next-web/src/middleware/database/mod.rs` 为 `next-web-actuator/src/actuate/mod.rs`
- 修改 `next-web/src/extract/find_singleton.rs` 中的导入语句
- 修改 `next-web/src/lib.rs` 中的导入语句以包含新的宏和模块
- 修改 `next-web/src/configurer/http_method_handler_configurer.rs` 以支持 OpenAPI Router
- 修改 `next-web/src/application/application.rs` 以支持 API 文档和中间件初始化
- 修改 `next-web/src/middleware/mod.rs` 删除不再需要的模块导入
- 修改 `next-web/src/middleware/redis/mod.rs` 删除 `redis_lock` 模块
- 修改 `next-web/src/stream/bytes_stream.rs` 优化响应头生成
- 修改 `next-web/src/util/digester.rs` 优化导入顺序和注释
- 修改 `next-web/src/util/thread.rs` 添加空行以符合代码规范
- 修改 `next-web-document/Cargo.toml` 移除未使用的特性
- 修改 `next-web-document/next-web-document-excel/fast-excel/src/core/poi/ss/usermodel/cell.rs` 中的导入语句
- 修改 `next-web-document/next-web-document-excel/fast-excel/src/core/poi/ss/formula/constant/constant_value_parser.rs` 以处理字符串解析错误
- 修改 `next-web-document/next-web-document-excel/fast-excel/src/core/util/work_book_util.rs` 添加创建行的方法
- 修改 `next-web-document/next-web-document-excel/fast-excel/src/core/util/write_handler_utils.rs` 添加创建行的方法
- 修改 `next-web-document/next-web-document-excel/fast-excel/src/core/excel_reader.rs` 添加 ExcelReader 实现
- 修改 `next-web-document/next-web-document-excel/fast-excel/src/core/excel_writer.rs` 添加 writer 参数
- 修改 `next-web-document/next-web-document-excel/fast-excel/src/core/write/builder/excel_writer_builder.rs` 支持 writer 参数
- 修改 `next-web-document/next-web-document-excel/fast-excel/src/core/write/metadata/mod.rs` 添加 row_data 和 base_write_holder
- 修改 `next-web-document/next-web-document-excel/fast-excel/src/core/write/metadata/write_sheet.rs` 支持 sheet_no 和 sheet_name 的可选性
- 修改 `next-web-document/next-web-document-excel/fast-excel/src/core/write/metadata/write_table.rs` 支持 table_no 的可选性
- 修改 `next-web-document/next-web-document-excel/fast-excel/src/core/event/mod.rs` 添加 analysis_event_listener 和 sync_read_listener
- 修改 `next-web-document/next-web-document-excel/fast-excel/src/core/event/analysis_event_listener.rs` 添加 AnalysisEventListener 实现
- 修改 `next-web-document/next-web-document-excel/fast-excel/src/core/event/sync_read_listener.rs` 添加 SyncReadListener 实现
- 修改 `next-web-macros/src/web/api_doc.rs` 实现 api_doc 宏
- 修改 `next-web-macros/src/web/mod.rs` 添加 api_doc 模块
- 修改 `next-web-macros/src/web/routing.rs` 优化路由注解处理和添加 API 文档集成
- 修改 `next-web-macros/src/lib.rs` 优化宏定义导入顺序和添加 api_doc 宏
- 修改 `next-web-tests/Cargo.toml` 添加 utoipa 依赖以支持 API 文档
- 修改 `next-web-tests/examples/test_api_doc.rs` 添加示例 API 文档集成
- 修改 `next-web-tests/examples/test_application_event.rs` 优化中间件初始化方法签名
- 修改 `PLAN.md` 添加 API 文档集成计划

Closes #123

+++ next-web-document/next-web-document-excel/fast-excel/src/core/poi/ss/util/cell_reference.rs
@@ -190,9 +190,9 @@ impl CellReference {
             }
 
             // Character is uppercase letter, find relative value to A
-            retval = (retval * 26) + (*ch as i32 - 'A' as i32 + 1);
+            retval = (retval * 26) + (*ch as u32 - 'A' as u32 + 1);
         }
-        retval - 1
+        (retval - 1) as i32
     }
 
     /// Classifies an identifier as either a simple (2D) cell reference or a named range name
